pyproject.toml
[project]
name = "notebooklm-study-guide-automator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "ffmpeg>=1.4",
    "google-generativeai>=0.8.5",
    "html2text>=2025.4.15",
    "manim>=0.19.0",
    "playwright>=1.54.0",
    "markdown",
    "python-dotenv",
    "jsonschema",
    "manim_voiceover",
]

[project.optional-dependencies]
test = [
    "pytest",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["gm*"] 
